finished module

+++ lab-kyle/index.js
@@ -2,10 +2,10 @@
 
 const fs = require('fs');
 const bitmap = fs.readFileSync('../asset/bitmap.bmp');
-const colorShift = require('./lib/color-shift.js'); 
+const colorShift = require('./lib/color-shift.js');
 
 console.log('Bitmap Type: ' + bitmap.toString('ascii', 0, 2));
-console.log('File Size: ' + bitmap.readUInt32LE(2));
+console.log('File Size: ' + bitmap.readUInt16LE(2));
 console.log('Offset Start: ' + bitmap.readUInt32LE(10));
 console.log('Bits Per Pxl: ' + bitmap.readUInt32LE(28));
 console.log('Number of colors: ' + bitmap.readUInt32LE(46));
